Reactify Reports component
