+++ src/components/reports/_reports.scss
@@ -1,0 +1,4 @@
+.sankey-box {
+  width: 100%;
+  height: 95vh;
+}
