--- src/components/reports/cost-by-service.njk
@@ -1,136 +1,0 @@
-{% extends "../../layouts/govuk.njk" %}
-
-{% block pageTitle %}
-  Statement
-{% endblock %}
-
-{% block content %}
-  {{ super() }}
-
-  <h1 class="govuk-heading-l">
-    Billables by service for {{ date }}
-  </h1>
-
-  <table class="govuk-table">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">
-          Service
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Including VAT
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Excluding VAT
-        </th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for record in billablesByService %}
-        <tr class="govuk-table__row">
-          <td class="govuk-table__cell">
-            {{ record.serviceGroup }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.incVAT | currency(2) }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.exVAT | currency(2) }}
-          </td>
-        </tr>
-      {% endfor %}
-    </tbody>
-    <tfoot>
-  </table>
-
-  <h1 class="govuk-heading-l">
-    Billables by organisation and service for {{ date }}
-  </h1>
-
-  <table class="govuk-table">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">
-          Organization
-        </th>
-        <th class="govuk-table__header" scope="col">
-          Service
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Including VAT
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Excluding VAT
-        </th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for record in billablesByOrganisationAndService %}
-        <tr class="govuk-table__row">
-          <td class="govuk-table__cell" scope="row">
-            {{ record.orgName }}
-          </td>
-          <td class="govuk-table__cell">
-            {{ record.serviceGroup }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.incVAT | currency(2) }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.exVAT | currency(2) }}
-          </td>
-        </tr>
-      {% endfor %}
-    </tbody>
-    <tfoot>
-  </table>
-
-  <h1 class="govuk-heading-l">
-    Billables by organisation and space and service for {{ date }}
-  </h1>
-
-  <table class="govuk-table">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">
-          Organization
-        </th>
-        <th class="govuk-table__header" scope="col">
-          Space
-        </th>
-        <th class="govuk-table__header" scope="col">
-          Service
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Including VAT
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Excluding VAT
-        </th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for record in billablesByOrganisationAndSpaceAndService %}
-        <tr class="govuk-table__row">
-          <td class="govuk-table__cell" scope="row">
-            {{ record.orgName }}
-          </td>
-          <td class="govuk-table__cell">
-            {{ record.spaceName }}
-          </td>
-          <td class="govuk-table__cell">
-            {{ record.serviceGroup }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.incVAT | currency(2) }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.exVAT | currency(2) }}
-          </td>
-        </tr>
-      {% endfor %}
-    </tbody>
-    <tfoot>
-  </table>
-
-{% endblock %}
