--- src/components/reports/cost.njk
@@ -1,112 +1,0 @@
-{% extends "../../layouts/govuk.njk" %}
-
-{% block pageTitle %}
-  Statement
-{% endblock %}
-
-{% block content %}
-  {{ super() }}
-
-  <h1 class="govuk-heading-l">
-    Billables for {{ date }}
-  </h1>
-
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-one-quarter">
-      <h2 class="govuk-heading-m">
-        {{ billableEventCount | dump}}
-        <span class="govuk-caption-m">Billable events</span>
-      </h2>
-    </div>
-    <div class="govuk-grid-column-one-quarter">
-      <h2 class="govuk-heading-m">
-        £{{ totalBillables.incVAT | currency(2) }}
-        <span class="govuk-caption-m">Total including VAT</span>
-      </h2>
-    </div>
-    <div class="govuk-grid-column-one-quarter">
-      <h2 class="govuk-heading-m">
-        £{{ totalBillables.exVAT | currency(2) }}
-        <span class="govuk-caption-m">Total excluding VAT</span>
-      </h2>
-    </div>
-  </div>
-
-  <h1 class="govuk-heading-l">
-    Billables by organisation for {{ date }}
-  </h1>
-
-  <table class="govuk-table">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">
-          Organization
-        </th>
-        <th class="govuk-table__header" scope="col">
-          Quota
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Including VAT
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Excluding VAT
-        </th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for record in orgCostRecords %}
-        <tr class="govuk-table__row">
-          <th class="govuk-table__header" scope="row">
-            {{ record.orgName }}
-          </th>
-          <td class="govuk-table__cell">
-            {{ record.quotaName }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.incVAT | currency(2) }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.exVAT | currency(2) }}
-          </td>
-        </tr>
-      {% endfor %}
-    </tbody>
-    <tfoot>
-  </table>
-
-  <h1 class="govuk-heading-l">
-    Billables by quota for {{ date }}
-  </h1>
-
-  <table class="govuk-table">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">
-          Quota
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Including VAT
-        </th>
-        <th class="govuk-table__header govuk-table__header--numeric" scope="col">
-          Excluding VAT
-        </th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for record in quotaCostRecords %}
-        <tr class="govuk-table__row">
-          <th class="govuk-table__header" scope="row">
-            {{ record.quotaName }}
-          </th>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.incVAT | currency(2) }}
-          </td>
-          <td class="govuk-table__cell govuk-table__cell--numeric">
-            £{{ record.exVAT | currency(2) }}
-          </td>
-        </tr>
-      {% endfor %}
-    </tbody>
-  </table>
-
-{% endblock %}
