+++ src/components/reports/index.ts
@@ -1,5 +1,1 @@
-export * from './cost';
-export * from './cost-by-service';
-export * from './pmo-org-spend';
-export * from './organizations';
-export * from './visualisation';
+export * from './controllers';
