--- src/components/reports/organizations.njk
@@ -1,119 +1,0 @@
-{% extends "../../layouts/govuk.njk" %}
-{% from "govuk-frontend/govuk/components/input/macro.njk" import govukInput %}
-{% from "govuk-frontend/govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk-frontend/govuk/components/panel/macro.njk" import govukPanel %}
-
-{% block pageTitle %}
-  Organisations
-{% endblock %}
-
-{% block content %}
-  {{ super() }}
-
-  <h1 class="govuk-heading-l">Organisations</h1>
-
-  <p class="govuk-body">
-    There
-    {{ organizations.length | plural('is', 'are') }}
-    {{ organizations.length }}
-    {{ organizations.length | plural('organisation', 'organisations') }}.
-  </p>
-
-  <h2 class="govuk-heading-m">Trial accounts</h2>
-
-  <p class="govuk-body">
-    There
-    {{ trialOrgs.length | plural('is', 'are') }}
-    {{ trialOrgs.length }}
-    trial {{ trialOrgs.length | plural('organisation', 'organisations') }}.
-  </p>
-
-  <p class="govuk-body">
-    Sorted by age; oldest first.
-  </p>
-
-  <table class="govuk-table">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header">Organisation</th>
-        <th class="govuk-table__header">Owner</th>
-        <th class="govuk-table__header">Quota</th>
-        <th class="govuk-table__header">Creation date</th>
-        <th class="govuk-table__header">Status</th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-    {% for organization in trialOrgs %}
-      <tr class="govuk-table__row">
-        <td class="govuk-table__cell">
-          <a href="{{linkTo('admin.organizations.view', {organizationGUID: organization.guid})}}" class="govuk-link">
-            {{ organization.name }}
-          </a>
-        </td>
-        <td class="govuk-table__cell">
-          {{organization.metadata.annotations.owner | default('Unknown')}}
-        </td>
-        <td class="govuk-table__cell">
-          {{orgQuotaMapping[organization.relationships.quota.data.guid].entity.name}}
-        </td>
-        <td class="govuk-table__cell">
-          {{organization.created_at | nicedate}}
-        </td>
-        <td class="govuk-table__cell">
-					{{orgTrialExpirys[organization.guid] | pastorfuture('Expired', 'Expires') -}}
-          {{- ' ' -}}
-          {{orgTrialExpirys[organization.guid] | relativetime}}
-        </td>
-      </tr>
-    {% endfor %}
-    </tbody>
-  </table>
-
-  <h2 class="govuk-heading-m">Billable accounts</h2>
-
-  <p class="govuk-body">
-    There
-    {{ billableOrgs.length | plural('is', 'are') }}
-    {{ billableOrgs.length }}
-    billable {{ billable.length | plural('organisation', 'organisations') }}.
-  </p>
-
-  <p class="govuk-body">
-    Sorted by age; newest first.
-  </p>
-
-  <table class="govuk-table">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header">Organisation</th>
-        <th class="govuk-table__header">Owner</th>
-        <th class="govuk-table__header">Quota</th>
-        <th class="govuk-table__header">Creation date</th>
-        <th class="govuk-table__header">Time since creation</th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-    {% for organization in billableOrgs %}
-      <tr class="govuk-table__row">
-        <td class="govuk-table__cell">
-          <a href="{{linkTo('admin.organizations.view', {organizationGUID: organization.guid})}}" class="govuk-link">
-            {{ organization.name }}
-          </a>
-        </td>
-        <td class="govuk-table__cell">
-          {{organization.metadata.annotations.owner | default('Unknown')}}
-        </td>
-        <td class="govuk-table__cell">
-          {{orgQuotaMapping[organization.relationships.quota.data.guid].entity.name}}
-        </td>
-        <td class="govuk-table__cell">
-          {{organization.created_at | nicedate}}
-        </td>
-        <td class="govuk-table__cell">
-          Created {{organization.created_at | relativetime}}
-        </td>
-      </tr>
-    {% endfor %}
-    </tbody>
-  </table>
-{% endblock %}
