--- src/components/reports/visualisation.njk
@@ -1,26 +1,0 @@
-{% extends "../../layouts/govuk.njk" %}
-
-{% block pageTitle %}
-  Statement
-{% endblock %}
-
-{% block content %}
-  {{ super() }}
-
-  <h1 class="govuk-heading-l">
-    Billing flow for {{ date }}
-  </h1>
-
-  {% if data %}
-  <script id="data" type="application/json">
-  {{data | dump | safe}}
-  </script>
-  <svg id="sankey" style="width: 100%; height: 95vh;"></svg>
-
-  <script src="/assets/d3.min.js"></script>
-  <script src="/assets/d3-sankey.min.js"></script>
-  <script src="/assets/sankey.js"></script>
-  {% else %}
-  <h2 class="govuk-heading-m">No data for {{ date }}</h2>
-  {% endif %}
-{% endblock %}
