+++ src/layouts/govuk.screen.scss
@@ -9,6 +9,7 @@ $govuk-page-width: 1200px;
 @import '../components/calculator/calculator';
 
 @import '../components/reports/subnav/element';
+@import '../components/reports/reports';
 
 @import '../components/org-users/org-users';
 @import '../components/org-users/permissions';
